+++ app/controllers/admin/UserController.php
@@ -7,8 +7,14 @@
 namespace SoftnCMS\controllers\admin;
 
 use SoftnCMS\controllers\BaseController;
+use SoftnCMS\controllers\Form;
 use SoftnCMS\controllers\Messages;
 use SoftnCMS\controllers\Pagination;
+use SoftnCMS\controllers\Router;
+use SoftnCMS\controllers\Token;
+use SoftnCMS\Helpers\ArrayHelp;
+use SoftnCMS\Helpers\Helps;
+use SoftnCMS\models\admin\template\Template;
 use SoftnCMS\models\admin\User;
 use SoftnCMS\models\admin\Users;
 use SoftnCMS\models\admin\UserInsert;
@@ -17,94 +23,110 @@ use SoftnCMS\models\admin\UserUpdate;
 
 /**
  * Clase del controlador de la pagina de usuarios.
- *
  * @author Nicolás Marulanda P.
  */
 class UserController extends BaseController {
-
+    
     /**
      * Metodo llamado por la funcion INDEX.
-     * @param array $data Lista de argumentos..
+     *
+     * @param array $data Lista de argumentos.
+     *
      * @return array
      */
     protected function dataIndex($data) {
-        $output = [];
-        $countData = Users::count();
-        $pagination = new Pagination($data['paged'], $countData);
-        $limit = $pagination->getBeginRow() . ',' . $pagination->getRowCount();
-        $users = Users::selectByLimit($limit);
-
-        $pagination->concatUrl('admin/user');
-
+        $output     = [];
+        $countData  = Users::count();
+        $pagination = new Pagination(ArrayHelp::get($data, 'paged'), $countData);
+        $limit      = $pagination->getBeginRow() . ',' . $pagination->getRowCount();
+        $users      = Users::selectByLimit($limit);
+        Template::setPagination($pagination);
+        
         if ($users !== \FALSE) {
             $output = $users->getAll();
         }
-
+        
         return [
             'users' => $output,
-            'pagination' => $pagination
         ];
     }
-
+    
     /**
      * Metodo llamado por la función INSERT.
      * @return array
      */
     protected function dataInsert() {
-        global $urlSite;
-
-        if (filter_input(\INPUT_POST, 'publish')) {
-
+        if (Form::submit('publish')) {
             $dataInput = $this->getDataInput();
-            if ($dataInput['userPass'] == $dataInput['userPassR']) {
+            
+            if ($dataInput !== FALSE && $dataInput['userPass'] == $dataInput['userPassR']) {
                 $insert = new UserInsert($dataInput['userLogin'], $dataInput['userName'], $dataInput['userEmail'], $dataInput['userPass'], $dataInput['userRol'], $dataInput['userUrl']);
-
+                
                 if ($insert->insert()) {
                     Messages::addSuccess('Usuario registrado correctamente.');
-                    //Si todo es correcto se muestra el USER en la pagina de edición.
-                    header("Location: $urlSite" . 'admin/user/update/' . $insert->getLastInsertId());
-                    exit();
+                    //Si es correcto se muestra el USER en la pagina de edición.
+                    Helps::redirectRoute('update/' . $insert->getLastInsertId());
                 }
-
-                Messages::addError('Error al registrar el usuario.');
             }
+            Messages::addError('Error al registrar el usuario.');
         }
-
+        
         return [
             //Datos por defecto a mostrar en el formulario.
             'user' => User::defaultInstance(),
-            /*
-             * Booleano que indica si muestra el encabezado
-             * "Agregar nuevo usuario" si es FALSE 
-             * o "Actualizar usuario" si es TRUE
-             */
-            'actionUpdate' => \FALSE,
         ];
     }
-
+    
+    /**
+     * Metodo que obtiene los datos de los campos INPUT del formulario.
+     * @return array|bool
+     */
+    protected function getDataInput() {
+        if (Token::check()) {
+            /*
+             * Si "GetDataInput" es llamado desde la función "insert"
+             * sera obligatorio los campos de las contraseñas.
+             */
+            $isInsert = Router::getRequest()->getMethod() == 'insert';
+            
+            Form::addInputAlphanumeric('userLogin', TRUE, FALSE, FALSE, FALSE, 1, TRUE, '');
+            Form::addInputAlphabetic('userName', TRUE);
+            Form::addInputEmail('userEmail', TRUE);
+            Form::addInputAlphanumeric('userPass', $isInsert);
+            Form::addInputAlphanumeric('userPassR', $isInsert);
+            Form::addInputInteger('userRol');
+            Form::addInputUrl('userUrl');
+            
+            return Form::postInput();
+        }
+        
+        return FALSE;
+    }
+    
     /**
      * Metodo llamado por la función UPDATE.
+     *
      * @param array $data Lista de argumentos.
+     *
      * @return array
      */
     protected function dataUpdate($data) {
-        global $urlSite;
-
-        $user = User::selectByID($data['id']);
-
+        $user = User::selectByID(ArrayHelp::get($data, 'id'));
+        
         //En caso de que no exista.
         if (empty($user)) {
             Messages::addError('Error. El usuario no existe.');
-            header("Location: $urlSite" . 'admin/user');
-            exit();
+            Helps::redirectRoute();
         }
-
-        if (filter_input(\INPUT_POST, 'update')) {
+        
+        if (Form::submit('update')) {
             $dataInput = $this->getDataInput();
-
-            if ($dataInput['userPass'] == $dataInput['userPassR']) {
+            
+            if ($dataInput === FALSE || $dataInput['userPass'] != $dataInput['userPassR']) {
+                Messages::addError('Error al actualizar el usuario.');
+            } else {
                 $update = new UserUpdate($user, $dataInput['userLogin'], $dataInput['userName'], $dataInput['userEmail'], $dataInput['userPass'], $dataInput['userRol'], $dataInput['userUrl']);
-
+                
                 //Si ocurre un error la función "$update->update()" retorna FALSE.
                 if ($update->update()) {
                     Messages::addSuccess('Usuario actualizado correctamente.');
@@ -114,23 +136,17 @@ class UserController extends BaseController {
                 }
             }
         }
-
+        
         return [
             //Instancia USER
             'user' => $user,
-            /*
-             * Booleano que indica si muestra el encabezado
-             * "Agregar nuevo usuario" si es FALSE 
-             * o "Actualizar usuario" si es TRUE
-             */
-            'actionUpdate' => \TRUE,
         ];
     }
-
+    
     /**
      * Metodo llamado por la función DELETE.
+     *
      * @param array $data Lista de argumentos.
-     * @return array
      */
     protected function dataDelete($data) {
         /*
@@ -138,10 +154,14 @@ class UserController extends BaseController {
          * se carga el modulo vista INDEX, asi que se retornan los datos
          * para esta vista.
          */
-
-        $delete = new UserDelete($data['id']);
-        $output = $delete->delete();
-
+        
+        $output = FALSE;
+        
+        if (Token::check()) {
+            $delete = new UserDelete($data['id']);
+            $output = $delete->delete();
+        }
+        
         if ($output) {
             Messages::addSuccess('Usuario borrado correctamente.');
         } elseif ($output === 0) {
@@ -149,24 +169,6 @@ class UserController extends BaseController {
         } else {
             Messages::addError('Error al borrar el usuario.');
         }
-
-        $this->namePage = 'user';
     }
-
-    /**
-     * Metodo que obtiene los datos de los campos INPUT del formulario.
-     * @return array
-     */
-    protected function getDataInput() {
-        return [
-            'userLogin' => \filter_input(\INPUT_POST, 'userLogin'),
-            'userName' => \filter_input(\INPUT_POST, 'userName'),
-            'userEmail' => \filter_input(\INPUT_POST, 'userEmail'),
-            'userPass' => \filter_input(\INPUT_POST, 'userPass'),
-            'userPassR' => \filter_input(\INPUT_POST, 'userPassR'),
-            'userRol' => \filter_input(\INPUT_POST, 'userRol'),
-            'userUrl' => \filter_input(\INPUT_POST, 'userUrl'),
-        ];
-    }
-
+    
 }
